store class names in redis
rails lazy loads worker classes, and so they may not exist. this stores
the list of worker classes centrally and will not matter if it has not
been loaded yet. while it requires an additional `sadd` on each stat
reported, it is worth the overhead and a requirement for any server
that becomes restarted or has multiple app servers.